requirements.txt
# Flask核心框架及扩展
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-WTF==1.1.1

# 数据序列化与验证
marshmallow==3.20.1

# 任务队列
redis==4.6.0
rq==1.15.1

# 环境配置
python-dotenv==1.0.0

# HTTP请求
requests==2.31.0

# AI服务相关依赖 - 统一使用Langchain集成
langchain>=0.2.0
langchain-community>=0.2.0
langchain_openai>=0.1.0
langgraph>=0.1.0
langsmith>=0.1.0
dashscope>=1.14.1

# 阿里云文档智能服务
alibabacloud_tea_openapi>=0.3.0
alibabacloud_docmind_api20220711==1.4.7
alibabacloud_credentials>=0.3.0

# 向量数据库
weaviate-client==3.25.3

# 数据处理
pandas>=1.5.0
numpy>=1.24.0

# 中文文本处理
jieba>=0.42.1

# 机器学习（可选，用于本地模型）
transformers>=4.30.0
torch>=2.0.0

# 评估工具（可选）
ragas>=0.1.0

# 开发工具
black>=23.0.0

# 测试框架
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
factory-boy>=3.3.0

# 其他工具库
werkzeug>=2.3.0
psutil>=5.9.0
